+++ lib/generators/happy_seed/splash/templates/app/controllers/splash_controller.rb
@@ -1,5 +1,5 @@
 class SplashController < ApplicationController
-  skip_before_filter :authenticate
+  skip_before_action :authenticate
 
   def index
   end