+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+Version 0.0.21
+  - Upgrade to Rails 5.0.0.beta3
+  - Upgrade rspec-rails to 3.5.0.beta1
+
+  
 Version 0.0.20
   - Updated Bootstrap panels in the templates
   - Disabled many default test generators